Removed OSCtoolset files from repo, removed other ue plugins from repo
removed paid plugin files from the repo. UE plugin files were moved to engine plugin folder so they still work in project if you also have the plugins, but vr tracking TD project had to have the tox completely removed for now.  Trying to figure out a way to keep the project intact if you have the paid plugin too, but not sure if an ext-tox reference will solve that